requirements.txt
Cython==0.23.1
Django==1.8.4
MySQL-python==1.2.5
argparse==1.2.1
django-tastypie==0.12.2
h5py==2.5.0
neo==0.3.3
numpy==1.9.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
quantities==0.10.1
requests==2.7.0
scipy==0.16.0
six==1.9.0
wsgiref==0.1.2
